build: fix swiftlint and swiftformat errors
Fix build errors and removes bootstrap that is not needed
