+++ .github/workflows/megalinter.yml
@@ -22,7 +22,7 @@ jobs:
           VALIDATE_ALL_CODEBASE: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
           DISABLE: COPYPASTE,SPELL
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          DISABLE_LINTERS: SWIFT_SWIFTLINT
+          DISABLE_LINTERS: SWIFT_SWIFTLINT,YAML_V8R,REPOSITORY_DEVSKIM,REPOSITORY_TRIVY
 
       - name: SwiftLint
         uses: norio-nomura/action-swiftlint@3.2.1
