--- .github/workflows/test.yml
@@ -1,37 +1,0 @@
----
-name: Test
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches: [main]
-
-jobs:
-  lint:
-    name: test
-    runs-on: macos-latest
-    steps:
-      - name: Checkout Code
-        uses: actions/checkout@v3
-
-      - name: Install SSH Key
-        uses: shimataro/ssh-key-action@v2.3.0
-        with:
-          key: ${{ secrets.SSH_PRIVATE_KEY }}
-          known_hosts: github.com
-
-      - name: ABC Install
-        uses: lyricwulf/abc@v1
-        with:
-          macos: openapi-generator
-
-      - name: Bootstrap
-        run: ./bootstrap.sh
-        env:
-          GITHUB_ACTOR: ${{ github.actor }}
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Test
-        run: "swift test"
