+++ .mega-linter.yml
@@ -1,0 +1,12 @@
+# Configuration file for MegaLinter
+# See all available variables at https://megalinter.io/configuration/ and in linters documentation
+DISABLE:
+  - COPYPASTE # Uncomment to disable checks of excessive copy-pastes
+  - SPELL # Uncomment to disable checks of spelling mistakes
+DISABLE_LINTERS:
+  - SWIFT_LINT
+  - YAML_V8R
+  - REPOSITORY_DEVSKIM
+SHOW_ELAPSED_TIME: true
+EXCLUDED_DIRECTORIES:
+  - PrismAPISDK
