+++ Apollo/Sources/Operations/RandomMnemonicsOperation.swift
@@ -1,11 +1,11 @@
+import Core
 import Foundation
 import PrismAPI
-import Core
 
 struct RandomMnemonicsOperation {
     let keyDerivation = KeyDerivation()
     let logger: PrismLogger
-    
+
     func compute() -> [String] {
         keyDerivation.randomMnemonicCode().words
     }
