+++ Authenticate/Tests/AuthenticateTests.swift
@@ -1,0 +1,3 @@
+import XCTest
+
+final class AuthenticateTests: XCTestCase {}
