--- Castor/Sources/DIDGrammar/DIDAbnfBaseListener.swift
@@ -1,89 +1,0 @@
-// Generated from java-escape by ANTLR 4.11.1
-
-import Antlr4
-
-
-/**
- * This class provides an empty implementation of {@link DIDAbnfListener},
- * which can be extended to create a listener which only needs to handle a subset
- * of the available methods.
- */
-open class DIDAbnfBaseListener: DIDAbnfListener {
-     public init() { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterDid(_ ctx: DIDAbnfParser.DidContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitDid(_ ctx: DIDAbnfParser.DidContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterMethod_name(_ ctx: DIDAbnfParser.Method_nameContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitMethod_name(_ ctx: DIDAbnfParser.Method_nameContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterMethod_specific_id(_ ctx: DIDAbnfParser.Method_specific_idContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitMethod_specific_id(_ ctx: DIDAbnfParser.Method_specific_idContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterIdchar(_ ctx: DIDAbnfParser.IdcharContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitIdchar(_ ctx: DIDAbnfParser.IdcharContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterEveryRule(_ ctx: ParserRuleContext) throws { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitEveryRule(_ ctx: ParserRuleContext) throws { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func visitTerminal(_ node: TerminalNode) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func visitErrorNode(_ node: ErrorNode) { }
-}
