--- Castor/Sources/DIDGrammar/DIDUrlAbnfBaseListener.swift
@@ -1,167 +1,0 @@
-// Generated from java-escape by ANTLR 4.11.1
-
-import Antlr4
-
-
-/**
- * This class provides an empty implementation of {@link DIDUrlAbnfListener},
- * which can be extended to create a listener which only needs to handle a subset
- * of the available methods.
- */
-open class DIDUrlAbnfBaseListener: DIDUrlAbnfListener {
-     public init() { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterDid_url(_ ctx: DIDUrlAbnfParser.Did_urlContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitDid_url(_ ctx: DIDUrlAbnfParser.Did_urlContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterDid(_ ctx: DIDUrlAbnfParser.DidContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitDid(_ ctx: DIDUrlAbnfParser.DidContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterMethod_name(_ ctx: DIDUrlAbnfParser.Method_nameContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitMethod_name(_ ctx: DIDUrlAbnfParser.Method_nameContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterMethod_specific_id(_ ctx: DIDUrlAbnfParser.Method_specific_idContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitMethod_specific_id(_ ctx: DIDUrlAbnfParser.Method_specific_idContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterPath(_ ctx: DIDUrlAbnfParser.PathContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitPath(_ ctx: DIDUrlAbnfParser.PathContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterQuery(_ ctx: DIDUrlAbnfParser.QueryContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitQuery(_ ctx: DIDUrlAbnfParser.QueryContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterFrag(_ ctx: DIDUrlAbnfParser.FragContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitFrag(_ ctx: DIDUrlAbnfParser.FragContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterSearch(_ ctx: DIDUrlAbnfParser.SearchContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitSearch(_ ctx: DIDUrlAbnfParser.SearchContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterSearchparameter(_ ctx: DIDUrlAbnfParser.SearchparameterContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitSearchparameter(_ ctx: DIDUrlAbnfParser.SearchparameterContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterString(_ ctx: DIDUrlAbnfParser.StringContext) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitString(_ ctx: DIDUrlAbnfParser.StringContext) { }
-
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func enterEveryRule(_ ctx: ParserRuleContext) throws { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func exitEveryRule(_ ctx: ParserRuleContext) throws { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func visitTerminal(_ node: TerminalNode) { }
-	/**
-	 * {@inheritDoc}
-	 *
-	 * <p>The default implementation does nothing.</p>
-	 */
-	open func visitErrorNode(_ node: ErrorNode) { }
-}
