+++ Domain/Sources/BBs/Apollo.swift
@@ -2,7 +2,7 @@ import Foundation
 
 public protocol Apollo {
     func createRandomMnemonics() -> [String]
-    func createSeed(mnemonics: [String]) throws -> Seed
+    func createSeed(mnemonics: [String], passphrase: String) throws -> Seed
     func createRandomSeed() -> (mnemonic: [String], seed: Seed)
     func createKeyPair(seed: Seed, index: Int) -> KeyPair
     func compressedPublicKey(publicKey: PublicKey) -> CompressedPublicKey
